refactor(a1): extract SeatGeek read routes to routers/seatgeek.py (BR-CODE-1 slice 9)
Moves the 6 read-only /api/seatgeek/* routes (event/listings/sales/
seller-listings/seller-orders/seller-status) out of app.py into
routers/seatgeek.py via the require_sb getter. All persisted-snapshot reads;
the link/sync POSTs stay in app.py. +5 tests. 86 tests green; app.py -171 LOC
(8423->8252).
